updated-sha-encryption-class
Added the source code and the respective referencing for the Class code, with only a minor name change by the programmer to better match the followed order as of now. Further source code to be added so the encryptors can be implemented to the facade and thus allow the user to encrypt messages.